refactor: move main layer mock dependencies to a test folder
