+++ src/main/test/mock-controller.ts
@@ -1,0 +1,18 @@
+import { ok } from '@/presentation/helpers/http/http-helper'
+import { Controller, HttpRequest, HttpResponse } from '@/presentation/protocols'
+
+const httpResponse: HttpResponse = ok({
+  id: 'any_id',
+  name: 'any_name',
+  email: 'any_email@email.com',
+  password: 'hashed_password'
+})
+
+export const mockController = (): Controller => {
+  class ControllerStub implements Controller {
+    async handle (httpRequest: HttpRequest): Promise<HttpResponse> {
+      return Promise.resolve(httpResponse)
+    }
+  }
+  return new ControllerStub()
+}
