+++ src/validation/test/index.ts
@@ -1,0 +1,1 @@
+export * from './mock-email-validator'
