+++ src/validation/test/mock-email-validator.ts
@@ -1,0 +1,10 @@
+import { EmailValidator } from '@/validation/protocols/email-validator'
+
+export const mockEmailValidator = (): EmailValidator => {
+  class EmailValidatorStub implements EmailValidator {
+    isValid (email: string): boolean {
+      return true
+    }
+  }
+  return new EmailValidatorStub()
+}
